Move components to seperate dirs
