+++ src/components/Countdown/index.js
@@ -1,0 +1,1 @@
+export { default } from './Countdown'
