+++ src/components/PageHeader/index.js
@@ -1,0 +1,1 @@
+export { default } from './PageHeader'
