+++ src/components/TroopsAtGoogle/TroopsAtGoogle.js
