+++ src/components/TroopsAtGoogle/index.js
@@ -1,0 +1,2 @@
+export { default } from './TroopsAtGoogle'
+export { default as TroopsAtGoogleImage } from './TroopsAtGoogle'
