+++ tests/__mocks__/gatsby.js
@@ -5,6 +5,7 @@ module.exports = {
   Link: () => 'gatsby-link',
   StaticQuery: () => 'static-query',
   useStaticQuery: args => {
+    // mock for useStaticQuery
     const query = args[0]
     if (query.indexOf('codeImage') > -1) {
       return {
